move track api calls to separate file
